remove metrics reporting from healthz
