+++ testgrid/tgapi/pkg/handlers/healthz.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/replicatedhq/kurl/testgrid/tgapi/pkg/persistence"
 	"github.com/replicatedhq/kurl/testgrid/tgapi/pkg/testinstance"
 )
 
@@ -17,18 +16,12 @@ func Healthz(w http.ResponseWriter, r *http.Request) {
 		IsAlive: true,
 	}
 
-	pending, running, timedOut, err := testinstance.GetTestStats()
+	err := testinstance.Healthz()
 	if err != nil {
-		fmt.Printf("error getting test stats: %s", err)
+		fmt.Printf("error connecting to the db: %s", err)
 		JSON(w, 500, healthzResponse)
-		return // being unable to get test stats indicates unhealthiness
+		return // being unable to connect to the db indicates unhealthiness
 	}
-
-	go func(pending, running, timedOut int64) {
-		persistence.MaybeSendStatsdGauge("pending", float64(pending), nil, 1.0)         // the number of pending test runs
-		persistence.MaybeSendStatsdGauge("running_3h", float64(running), nil, 1.0)      // the number of test runs that have been running for <3h
-		persistence.MaybeSendStatsdGauge("running_3h_24h", float64(timedOut), nil, 1.0) // the number of test runs that have been running for >3h and <24h (and that can be assumed to have timed out)
-	}(pending, running, timedOut)
 
 	JSON(w, 200, healthzResponse)
 }
