+++ testgrid/tgapi/pkg/testinstance/testinstance.go
@@ -546,3 +546,17 @@ func GetNodeStatus(id string) (string, error) {
 
 	return status.String, nil
 }
+
+func Healthz() error {
+	db := persistence.MustGetPGSession()
+
+	query := `select 1`
+	row := db.QueryRow(query)
+
+	var status int
+	if err := row.Scan(&status); err != nil {
+		return errors.Wrap(err, "failed to scan")
+	}
+
+	return nil
+}
